+++ src/Electron/VeloxMundiDesktop/src/pages/edit.html
@@ -5,7 +5,7 @@
     <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
     <meta
       http-equiv="Content-Security-Policy"
-      content="default-src 'self'; style-src 'self' 'unsafe-inline';"
+      content="default-src 'self'; style-src 'self' 'unsafe-inline'; img-src * http: https: data:"
     />
 
     <link rel="stylesheet" href="../scripts/resources/bootstrap-4.3.1/css/bootstrap.min.css" />
@@ -14,7 +14,7 @@
     <title>Velox Mundi: Edit</title>
   </head>
   <body>
-    <div class="row">
+    <div class="row" id="md-edit-toolbar">
       <div class="col-6">
         <button id="InsertImage" class="btn btn-default" title="Insert image"><span class="bi-image"></span></button>
       </div>
