+++ src/Electron/VeloxMundiDesktop/src/pages/options_prefs.html
@@ -10,29 +10,57 @@
 
     <link rel="stylesheet" href="../scripts/resources/bootstrap-4.3.1/css/bootstrap.min.css" />
     <link rel="stylesheet" href="../styles/custom.css" />
+    <link rel="stylesheet" href="../styles/options.css" />
     <title>Velox Mundi</title>
   </head>
   <body>
     <button class="btn btn-danger windowCloser bi-x" id="CloseOptions" title="Close">&nbsp;</button>
     <div class="titlebar">
       <h3><span class="bi-person-circle">&nbsp;</span>User Preferences</h3>
-    </div>
-    <a class="navLink nolink" href="#" data-page="options_main.html"><span class="bi-gear"></span>&nbsp;Back to Options</a>
-    <hr/>
-    <h3>Settings</h3>
+    </div><br/>
+    <h4 class="row" style="margin-left:0px; margin-right:0px; width:auto;">
+      <div class="col-10">
+        <a class="navLink nolink" href="#" data-page="options_main.html"><span class="bi-gear"></span>&nbsp;Back to Options</a>
+      </div>
+      <div class="col-2 text-right" style="padding-right:0px;">
+        <button id="SavePrefs" class="btn btn-success"><span class="bi-save">&nbsp;</span>Save</button>
+      </div>      
+    </h4>
     <table width="100%" class="configTable">
       <tr valign="top">
         <th style="width:20%;max-width:250px;">
-          Editor Style
+          
         </th>
         <td style="width:60%;max-width:100%;">
-          Choose whether to use a rich-text editor or markdown
+          
         </td>
         <td style="width:20%;max-width:250px;text-align:right;">
-          <select id="EditorStyle">
+          
+        </td>
+      </tr>
+      <tr valign="top">
+        <th>
+          Editor Style
+        </th>
+        <td>
+          Choose whether to use a rich-text editor or markdown
+        </td>
+        <td>
+          <select id="EditorStyle" class="form-control">
             <option value="RTE">Rich Text</option>
             <option value="MD">Markdown</option>
           </select>
+        </td>
+      </tr>
+      <tr>
+        <th>
+          Message Display
+        </th>
+        <td>
+          How many seconds should a message display on the page?
+        </td>
+        <td>
+          <input id="ToastTimeout" type="number" length="10" class="form-control"/>
         </td>
       </tr>
     </table>
