+++ src/Electron/VeloxMundiDesktop/src/styles/options.css
@@ -1,0 +1,3 @@
+#toast {
+  margin-top:40px;
+}
